BOOST_MESSAGE -> BOOST_TEST_MESSAGE which is the standard for newer boost libs; cast added in setAdaptivity in IterativeGridGeneratorSOO.cpp
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@5161 4eea3252-f0fb-4393-894d-40516dce545b

+++ optimization/src/sgpp/optimization/gridgen/IterativeGridGeneratorSOO.cpp
@@ -30,7 +30,7 @@ namespace SGPP {
 
     void IterativeGridGeneratorSOO::setAdaptivity(float_t adaptivity) {
       hMax = [adaptivity](size_t n) {
-        return std::pow(n, adaptivity);
+        return static_cast<size_t>(std::pow(n, adaptivity));
       };
     }
